+++ SoccerProject/src/main/java/com/itacademy/soccer/dao/IPlayerActionsDAO.java
@@ -17,8 +17,4 @@ public interface IPlayerActionsDAO extends JpaRepository<PlayerActions, PlayerMa
 
     List<PlayerActions> findByIdPlayerId (Long playerId);
 
-    @Override
-    default <S extends PlayerActions> S save(S playerActions) {
-        return save(playerActions);
-    }
 }
