+++ SoccerProject/src/main/java/com/itacademy/soccer/dao/IStadiumDAO.java
@@ -1,6 +1,7 @@
 package com.itacademy.soccer.dao;
 
 
+import com.itacademy.soccer.controller.json.StadiumJson;
 import com.itacademy.soccer.dto.Stadium;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -9,6 +10,5 @@ import org.springframework.stereotype.Repository;
 public interface IStadiumDAO extends JpaRepository<Stadium, Long> {
 
     Stadium findByStadiumId(Long stadiumId);
-
 
 }
