+++ SoccerProject/src/main/java/com/itacademy/soccer/service/IMatchService.java
@@ -2,9 +2,9 @@ package com.itacademy.soccer.service;
 
 import java.util.Date;
 import java.util.List;
-import java.util.Optional;
 
 import com.itacademy.soccer.dto.Match;
+import com.itacademy.soccer.dto.Player;
 
 public interface IMatchService {
 	
@@ -13,5 +13,6 @@ public interface IMatchService {
 	public List<Match> showAllMatches();
 	
 	public Match createMatch( Long localTeamId, Long visitorTeamId, Date date);
+
 
 }
