+++ SoccerProject/src/main/java/com/itacademy/soccer/service/IPlayerActionsService.java
@@ -1,26 +1,17 @@
 package com.itacademy.soccer.service;
 
-import com.itacademy.soccer.dao.IPlayerActionsDAO;
+import com.itacademy.soccer.dto.Match;
+import com.itacademy.soccer.dto.Player;
 import com.itacademy.soccer.dto.PlayerActions;
-import com.itacademy.soccer.dto.serializable.PlayerMatchId;
-import org.springframework.stereotype.Service;
 
+import java.util.HashMap;
 import java.util.List;
 
 
-public interface IPlayerActionsService extends IPlayerActionsDAO {
+public interface IPlayerActionsService  {
 
-
-	public PlayerActions findByIdPlayerIdAndIdMatchId(Long playerId, Long matchId);
-
-	List<PlayerActions> findByIdPlayerId(Long playerId);
-
-//	public PlayerActions updatePlayerActions(PlayerActions playerActionss);
-
-	//PlayerActions findByPlayerIdAndMatchId(Long playerId, Long matchId);
-
-
-
-	public PlayerActions save(PlayerActions playerActions);
-
+    void createActionsForPLayers(Match match, List<Player> playersList);
+    void initializeActionsInMatch(Match match, Player player);
+    int getOnePlayerActionsInOneMatch(PlayerActions playerActions, String action);
+    HashMap<String,Object> verifyIds(String ids, String id2, HashMap<String, Object> map);
 }
