+++ SoccerProject/src/main/java/com/itacademy/soccer/service/ITeamService.java
@@ -7,6 +7,7 @@ package com.itacademy.soccer.service;
 import java.util.List;
 
 import com.itacademy.soccer.dto.Team;
+import com.itacademy.soccer.service.impl.TeamServiceImpl;
 
 /**
  * @author KevHaes
@@ -14,6 +15,7 @@ import com.itacademy.soccer.dto.Team;
  */
 public interface ITeamService {
 
+	void createTeamInitial(TeamServiceImpl teamService);
 	public Team createTeam(Team team);
 
 	public List<Team> getAllTeams();
