+++ SoccerProject/src/main/java/com/itacademy/soccer/service/impl/PlayerActionsServiceImpl.java
@@ -1,32 +1,30 @@
 package com.itacademy.soccer.service.impl;
 
+import com.itacademy.soccer.dto.Match;
 import com.itacademy.soccer.dto.Player;
 import com.itacademy.soccer.dto.serializable.PlayerMatchId;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.data.domain.Example;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import com.itacademy.soccer.dao.IPlayerActionsDAO;
 import com.itacademy.soccer.dto.PlayerActions;
 import com.itacademy.soccer.service.IPlayerActionsService;
 
+import java.util.HashMap;
 import java.util.List;
-import java.util.Optional;
+
+import static com.itacademy.soccer.dto.lineup.Lineup.KEEPER;
 
 @Service
 public class PlayerActionsServiceImpl implements IPlayerActionsService{
 
-	@Qualifier("IPlayerActionsService")
+
 	@Autowired
 	IPlayerActionsDAO iPlayerActionsDAO;
 
 	//Este método no lo veo bien.
 	//find by playerMatchId
-	@Override
+
 	public PlayerActions findByIdPlayerIdAndIdMatchId(Long playerId, Long matchId) {
 
 		/*PlayerMatchId playerMatchId = iPlayerActionsDAO.findById(playerMatchId);
@@ -36,145 +34,76 @@ public class PlayerActionsServiceImpl implements IPlayerActionsService{
 		return iPlayerActionsDAO.findByIdPlayerIdAndIdMatchId(playerId, matchId);
 	}
 
-	@Override
 	public List<PlayerActions> findByIdPlayerId(Long playerId) {
 		return iPlayerActionsDAO.findByIdPlayerId(playerId);
 	}
 
-
-//	@Override
-//	public PlayerActions updatePlayerActions(PlayerActions playerActions) {
-//		return iPlayerActionsDAO.save(playerActions);
-//	}
-
-
-//	@Override
-//	public PlayerActions findByPlayerIdAndMatchId(Long playerId, Long matchId) {
-//		return iPlayerActionsDAO.findByPlayerIdAndMatchId(playerId, matchId);
-//	}
-
-
-
-	@Override
-	public PlayerActions save(PlayerActions playerActions) {
-		return iPlayerActionsDAO.save(playerActions);
-	}
-
-
-	@Override
-	public List<PlayerActions> findAll() {
-		return null;
-	}
-
-	@Override
-	public List<PlayerActions> findAll(Sort sort) {
-		return null;
-	}
-
-	@Override
-	public List<PlayerActions> findAllById(Iterable<PlayerMatchId> iterable) {
-		return null;
-	}
-
-	@Override
-	public <S extends PlayerActions> List<S> saveAll(Iterable<S> iterable) {
-		return null;
-	}
-
-	@Override
-	public void flush() {
+	public void createActionsForPLayers(Match match, List<Player> playersList){
+		try {
+			for (Player player : playersList) {
+				initializeActionsInMatch(match, player);
+				// guardar actions tambien en List de player.... inncesario
+//				player.getPlayerActions().add(actions);
+//				iPlayerDAO.save(player);
+				//	iPlayerActionsDAO.save(actions);
+			}
+		}catch (Exception e){
+			e.printStackTrace();
+			System.out.println(" Error in createActionsForPlayer ....");
+		}
 
 	}
+	public void initializeActionsInMatch(Match match, Player player){
 
-	@Override
-	public <S extends PlayerActions> S saveAndFlush(S s) {
-		return null;
+		PlayerMatchId playerMatchId = new PlayerMatchId();
+		playerMatchId.setPlayerId(player.getId());
+		playerMatchId.setMatchId(match.getId());
+
+		PlayerActions actions = new PlayerActions();
+		actions.setId(playerMatchId);
+		actions.setGoals(0);
+		actions.setAssists(0);
+		actions.setFouls(0);
+		actions.setRedCards(0);
+		actions.setYellowCards(0);
+		actions.setSaves(0);
+		actions.setLineup(null);
+		actions.setPlayer(player);
+		actions.setMatch(match);
+
+		try {
+			iPlayerActionsDAO.save(actions);
+		}catch (Exception e){
+			e.printStackTrace();
+		}
 	}
+	public int getOnePlayerActionsInOneMatch(PlayerActions playerActions, String action){
+		int data = 0;
 
-	@Override
-	public void deleteInBatch(Iterable<PlayerActions> iterable) {
-
+		if ( action.equalsIgnoreCase("goals")) data = playerActions.getGoals();
+		else
+		if ( action.equalsIgnoreCase("assists")) data = playerActions.getAssists();
+		else
+		if ( action.equalsIgnoreCase("fouls")) data = playerActions.getFouls();
+		else
+		if ( action.equalsIgnoreCase("red_cards")) data = playerActions.getRedCards();
+		else
+		if ( action.equalsIgnoreCase("yellow_cards")) data = playerActions.getYellowCards();
+		else
+		if ( action.equalsIgnoreCase("saves")) data = playerActions.getSaves();
+		else return -1;
+		return data;
 	}
+	public HashMap<String,Object> verifyIds(String ids, String id2, HashMap<String, Object> map){
+		Long id;
+		try{
+			id = Long.parseLong( ids );
 
-	@Override
-	public void deleteAllInBatch() {
-
+			if ( id2 != null) id = Long.parseLong( id2 );
+		}catch (Exception e){
+			map.put("success", false);
+			map.put("message", "Incorrect data Id  " );
+		}
+		return map;
 	}
-
-	@Override
-	public PlayerActions getOne(PlayerMatchId playerMatchId) {
-		return null;
-	}
-
-	@Override
-	public <S extends PlayerActions> List<S> findAll(Example<S> example) {
-		return null;
-	}
-
-	@Override
-	public <S extends PlayerActions> List<S> findAll(Example<S> example, Sort sort) {
-		return null;
-	}
-
-	@Override
-	public Page<PlayerActions> findAll(Pageable pageable) {
-		return null;
-	}
-
-	@Override
-	public Optional<PlayerActions> findById(PlayerMatchId playerMatchId) {
-		return Optional.empty();
-	}
-
-	@Override
-	public boolean existsById(PlayerMatchId playerMatchId) {
-		return false;
-	}
-
-	@Override
-	public long count() {
-		return 0;
-	}
-
-	@Override
-	public void deleteById(PlayerMatchId playerMatchId) {
-
-	}
-
-	@Override
-	public void delete(PlayerActions playerActions) {
-
-	}
-
-	@Override
-	public void deleteAll(Iterable<? extends PlayerActions> iterable) {
-
-	}
-
-	@Override
-	public void deleteAll() {
-
-	}
-
-	@Override
-	public <S extends PlayerActions> Optional<S> findOne(Example<S> example) {
-		return Optional.empty();
-	}
-
-	@Override
-	public <S extends PlayerActions> Page<S> findAll(Example<S> example, Pageable pageable) {
-		return null;
-	}
-
-	@Override
-	public <S extends PlayerActions> long count(Example<S> example) {
-		return 0;
-	}
-
-	@Override
-	public <S extends PlayerActions> boolean exists(Example<S> example) {
-		return false;
-	}
-
-
 }
