+++ SoccerProject/src/main/resources/application.properties
@@ -41,3 +41,5 @@ logging.level.org.hibernate.type=TRACE
 #spring.jpa.properties.hibernate.jdbc.time_zone=UTC
 spring.jpa.properties.hibernate.jdbc.time_zone=Europe/Madrid
 
+#####  Solventa ERROR al haber clases Serializadas y otras no que estan vinculadas con ManyToOne
+spring.jackson.serialization.fail-on-empty-beans=false
